+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added 
 - Support for autolinks (https://spec.commonmark.org/0.29/#autolinks)
   - Autolinks are rendered the same way as regular links.
+- Support for converting YouTube links to video preview
+  - This feature is disabled by defualt. Set `LinkStyle.LoadYouTubePreview = true` for enabling.
+  - There are also callbacks `LinkStyle.YouTubePreview.CustomLoadImage` for downloading image and `LinkStyle.YouTubePreview.TransformView` for rendering it
 
 ## 1.1.0-pre6 - 2021-02-24